backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5
pydantic-settings==2.7.1

# Supabase
supabase==2.11.0

# LLM via Fireworks (OpenAI-compatible) + LangGraph
openai==1.63.2
langchain==0.3.17
langchain-openai==0.3.5
langgraph==0.2.74

# Google Gemini — Veo 3.1 video, Nano Banana images, Flash chat + thinking + search
google-genai

# Utils
python-dotenv==1.0.1
httpx==0.28.1
